+++ lib/src/transports/websocket_dart_impl.dart
@@ -2,6 +2,8 @@ import 'dart:io';
 import 'dart:math';
 import 'dart:convert';
 import 'dart:async';
+import 'package:sip_ua/src/sip_ua_helper.dart';
+
 import '../logger.dart';
 import '../grammar.dart';
 
@@ -19,19 +21,17 @@ class WebSocketImpl {
   WebSocketImpl(this._url);
 
   void connect(
-      {Iterable<String> protocols,
-      Map<String, String> extHeaders,
-      bool allowBadCertificate = false}) async {
-    logger.info('connect $_url, $extHeaders, $protocols');
+      {Iterable<String> protocols, WebSocketSettings webSocketSettings}) async {
+    logger.info('connect $_url, ${webSocketSettings.extraHeaders}, $protocols');
     try {
-      if (allowBadCertificate) {
+      if (webSocketSettings.allowBadCertificate) {
         /// Allow self-signed certificate, for test only.
         var parsed_url = Grammar.parse(_url, 'absoluteURI');
-        _socket = await _connectForBadCertificate(
-            parsed_url.scheme, parsed_url.host, parsed_url.port, extHeaders);
+        _socket = await _connectForBadCertificate(parsed_url.scheme,
+            parsed_url.host, parsed_url.port, webSocketSettings);
       } else {
         _socket = await WebSocket.connect(_url,
-            protocols: protocols, headers: extHeaders);
+            protocols: protocols, headers: webSocketSettings.extraHeaders);
       }
 
       this?.onOpen();
@@ -62,12 +62,17 @@ class WebSocketImpl {
 
   /// For test only.
   Future<WebSocket> _connectForBadCertificate(String scheme, String host,
-      int port, Map<String, String> extHeaders) async {
+      int port, WebSocketSettings webSocketSettings) async {
     try {
       var r = new Random();
-      var key = base64.encode(List<int>.generate(8, (_) => r.nextInt(255)));
+      var key = base64.encode(List<int>.generate(16, (_) => r.nextInt(255)));
       var securityContext = new SecurityContext();
       var client = HttpClient(context: securityContext);
+
+      if (webSocketSettings.userAgent != null) {
+        client.userAgent = webSocketSettings.userAgent;
+      }
+
       client.badCertificateCallback =
           (X509Certificate cert, String host, int port) {
         logger.warn('Allow self-signed certificate => $host:$port. ');
@@ -76,18 +81,18 @@ class WebSocketImpl {
 
       var request = await client.getUrl(Uri.parse(
           (scheme == 'wss' ? 'https' : 'http') +
-              '://$host:$port/ws')); // form the correct url here
+              '://$host:$port')); // form the correct url here
+      request.headers.add('Connection', 'Upgrade', preserveHeaderCase: true);
+      request.headers.add('Upgrade', 'websocket', preserveHeaderCase: true);
+      request.headers.add('Sec-WebSocket-Version', '13',
+          preserveHeaderCase: true); // insert the correct version here
+      request.headers.add('Sec-WebSocket-Key', key.toLowerCase(),
+          preserveHeaderCase: true);
+      request.headers
+          .add('Sec-WebSocket-Protocol', 'sip', preserveHeaderCase: true);
 
-      request.headers.add('Connection', 'Upgrade');
-      request.headers.add('Upgrade', 'websocket');
-      request.headers.add(
-          'Sec-WebSocket-Version', '13'); // insert the correct version here
-      request.headers.add('Sec-WebSocket-Key', key.toLowerCase());
-
-      //request.headers.add('Origin', 'http://localhost:5060');
-      //request.headers.add('Sec-WebSocket-Protocol', 'sip');
-      extHeaders.forEach((key, value) {
-        request.headers.add(key, value);
+      webSocketSettings.extraHeaders.forEach((key, value) {
+        request.headers.add(key, value, preserveHeaderCase: true);
       });
 
       var response = await request.close();