+++ CHANGELOG.md
@@ -1,4 +1,22 @@
-## Changelog
+# Changelog
+
+[0.2.4] - 2020.08.25
+
+* Add missing key field `Sec-WebSocket-Protocol`.
+
+[0.2.3] - 2020.08.25
+
+* Add display_name for Call.
+* Add WebSocketSettings.
+* Fix the invalid extraHeaders in Registrator.
+* Exposed local_identity for Call.
+* Fixed Sec-WebSocket-Key keys are not 24 bytes.
+
+[0.2.2] - 2020.07.16
+
+* Refactor call API, move answer, hangup, hold etc methos to Call class.
+* Add SIP message listener to listen the new incoming SIP text message.
+* Expose ha1 in UaSettings.
 
 [0.2.1] - 2020.06.12